Rework the default handlers into layer-openstack-*
Due to commit 95493a4 in charms.reactive, the default handlers in
charms.openstack broke.  This is because charms.reactive no longer will
run handlers that are not in the hooks/ or reactive/ directory tree; the
default handlers in charms.openstack are in the library code, which is
pip installed into the module packages.

This patch, and related patches in charms.openstack and layer-openstack
enable the default handlers to function again.  Note that from a charm
author perspective, the API to default handlers is identical.  This is
merely an implementation change.

Change-Id: I38d46eaab3d64dd43166957a4aee01163bc00f5b
Partial-Bug: #1707685