Update HAProxy default timeout values
The default HAProxy values are set in the layer-openstack template for
haproxy. Bring the layer-openstack-api default configuration in line
with the classic charms which do not set values in the config.yaml but
rely on the template. If any future changes are required, a change in
one place (layer-openstack) will affect all charms.

Change-Id: I045d66bf606ef57b066b4d6af1d4829eeed3bc93